There are several things going on here. First, bool usegoodrun -> bool isdata, the end result of which is that the user no longer gets to decide  whether  or not to apply the goodrun selection to data, it is now always applied  to  data which means that the user _must_ initialize the goodrun tool etc. I can  imag- ine use cases for not doing this but having that option would  make  this  too general a tool and who wants that.
Second, aesthetics. Histograms are sorted by their Integral size before  being
stacked, the end result of which is that the histogram with the largest contr-
bution is on the top, the histogram with the second  largest  contribution  is
second from the top, etc. Also, we now use a modified tdrstyle among other lit-
tle changes I made which makes the plots pretty enough to eat.

Third, added duplicate event removal for data. This is only for data, so  must
trust that whatever MC is being used doesn't have duplicates. If you are para-
noid and want duplicate event removal in MC probably you will need  to  modify
the babies. Note that this is a hypothesis based duplicate event removal,  not
an event based duplicate event removal. Far be it for me to decide --  and  to
figure out in the first place -- which of several hypotheses in  an  event  to
keep in a generic way for every possible gathering selection...